Refactored movement to support context-sensitive moves
Changed some of the top-level movement functions to take a game object
instead of a color and a board. This will be needed to support things
like castling, en-passant, and stalemate rules that depend on past
moves.

I also simplified the moves stored in the game object to make things
easier at the expense of the quality of move printing. I plan to fix
this later since move printing already needed some improvements.